// src/features/add_employee/model/user.ts
export interface User {
    username: string
    password: string
    email: string
    lastName: string
    firstName: string
    dob: string
    phone: string
    gender: boolean
    address: string
    country: string
    citizenId: string
    departmentId: number
    isManager: boolean
    bankAccount: string
    bankAccountName: string
    bank: string
  }